% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM-Simulate.R
\name{GBM.Simulate}
\alias{GBM.Simulate}
\title{Simulate a Geometric Brownian Motion (GBM) Stochastic Process Through Monte Carlo Simulation}
\usage{
GBM.Simulate(n, t, mu, sigma, S0, dt)
}
\arguments{
\item{n}{The total number of price paths to simulate}

\item{t}{The forecasting period in years}

\item{mu}{The drift rate of the Geometric Brownian Motion process}

\item{sigma}{The volatility term of the Geometric Brownian Motion process}

\item{S0}{The initial value of the underlying asset}

\item{dt}{The discrete time step of observations, in years}
}
\value{
A matrix of simulated price paths of the GBM process. Each column corresponds to a simulated price path, and each
row corresponds to a simulated observed price of the simulated price paths at each discrete time period.
}
\description{
Simulate a Geometric Brownian Motion (GBM) stochastic process through Monte Carlo simulation. The GBM process is a commonly used stochastic
process to simulate the price paths of stock prices and other assets. The \code{GBM.Simulate} function utilizes antithetic variates as a
simple variance reduction technique.
}
\details{
A stochastic process S(t) is a geometric brownian motion that follows the following formula:
\deqn{ \frac{dS(t)}{S(t)} = \mu dt + \sigma dW(t)}{dS(t)/S(t) = mu dt + sigma dW(t)}

The GBM is log-normally distributed.
}
\examples{
## 100 simulations of 1 year of monthly price paths:
Simulated <- GBM.Simulate(n = 100,
                         t = 1,
                         mu = 0.05,
                         sigma = 0.2,
                         S0 = 100,
                         dt = 1/12)
}
