% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{LSM.AmericanOption}
\alias{LSM.AmericanOption}
\title{Value American-Style Options Through Least-Squares Monte Carlo (LSM) Simulation}
\usage{
LSM.AmericanOption(
  state.variables,
  payoff,
  K,
  dt,
  rf,
  call = FALSE,
  orthogonal = "Power",
  degree = 2,
  cross.product = TRUE,
  verbose = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{
# Price a vanilla American put option on an asset that follows
# Geometric Brownian Motion

## Step 1 - Simulate stock prices:
stock_prices <- GBM_simulate(n = 100, t = 1, mu = 0.05,
                         sigma = 0.2, S0 = 100, dt = 1/2)

## Step 2 - Value the American put option:

option_value <- LSM.AmericanOption(state.variables = stock_prices,
                                 payoff = stock_prices,
                                 K = 100,
                                 dt = 1/2,
                                 rf = 0.05)
# ->
option_value <- LSM_american_option(state_variables = stock_prices,
                                 payoff = stock_prices,
                                 K = 100,
                                 dt = 1/2,
                                 rf = 0.05)

}
\keyword{internal}
