% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{LSM.RealOption.OF}
\alias{LSM.RealOption.OF}
\title{Value operationally flexible capital investment projects through least-squares Monte Carlo (LSM) simulation:}
\usage{
LSM.RealOption.OF(
  state.variables,
  NCF,
  CAPEX,
  dt,
  rf,
  construction = 0,
  orthogonal = "Laguerre",
  degree = 9,
  cross.product = TRUE,
  Suspend.CAPEX = NULL,
  Suspend.OPEX = NULL,
  Resume.CAPEX = NULL,
  Abandon.CAPEX = NULL,
  Save.States = FALSE,
  verbose = FALSE,
  debugging = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

# Example: Value a capital investment project where the revenues follow a
# Geometric Brownian Motion stochastic process:

## Step 1 - Simulate asset prices:
asset_prices <- GBM_simulate(n = 100, t = 10, mu = 0.05,
                         sigma = 0.2, S0 = 100, dt = 1/2)

## Step 2 - Perform Real Option Analysis (ROA):


ROA <- LSM.RealOption.OF(state.variables = asset_prices,
                     NCF = asset_prices - 100,
                     CAPEX = 1000,
                     dt = 1/2,
                     rf = 0.05,
                     Suspend.CAPEX = 100,
                     Suspend.OPEX = 10,
                     Resume.CAPEX = 100,
                     Abandon.CAPEX = 0
                     )

# ->

ROA <- LSM_real_option_OF(state_variables = asset_prices,
                     NCF = asset_prices - 100,
                     CAPEX = 1000,
                     dt = 1/2,
                     rf = 0.05,
                     suspend_CAPEX = 100,
                     suspend_OPEX = 10,
                     resume_CAPEX = 100,
                     abandon_CAPEX = 0
                     )
}
\keyword{internal}
