% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{LSM.RealOption}
\alias{LSM.RealOption}
\title{Value capital investment projects through Least-Squares Monte Carlo (LSM) simulation:}
\usage{
LSM.RealOption(
  state.variables,
  NCF,
  CAPEX,
  dt,
  rf,
  construction = 0,
  orthogonal = "Laguerre",
  degree = 9,
  cross.product = TRUE,
  verbose = FALSE,
  debugging = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{
# Example: Value a capital investment project where the revenues follow a
# Geometric Brownian Motion stochastic process:

## Step 1 - Simulate asset prices:
AssetPrices <- GBM_simulate(n = 100, t = 10, mu = 0.05,
                         sigma = 0.2, S0 = 100, dt = 1/2)

## Step 2 - Perform Real Option Analysis (ROA):
ROA <- LSM.RealOption(state.variables = AssetPrices,
                     NCF = AssetPrices - 100,
                     CAPEX = 1000,
                     dt = 1/2,
                     rf = 0.05)
# ->

## Step 2 - Perform Real Option Analysis (ROA):
ROA <- LSM_real_option(state_variables = AssetPrices,
                     NCF = AssetPrices - 100,
                     CAPEX = 1000,
                     dt = 1/2,
                     rf = 0.05)

}
\keyword{internal}
