% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{GBM.Simulate}
\alias{GBM.Simulate}
\title{Simulate the geometric Brownian motion (GBM) stochastic process through Monte Carlo simulation}
\usage{
GBM.Simulate(n, t, mu, sigma, S0, dt)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{
Simulated <- GBM.Simulate(n = 100,
                         t = 1,
                         mu = 0.05,
                         sigma = 0.2,
                         S0 = 100,
                         dt = 1/12)

# ->

## 100 simulations of 1 year of monthly price paths:
Simulated <- GBM_simulate(n = 100,
                         t = 1,
                         mu = 0.05,
                         sigma = 0.2,
                         S0 = 100,
                         dt = 1/12)

}
\keyword{internal}
